Implement the definitions of Number, Add, and Multiply classes
We’ll start by looking at the semantics of SIMPLE expressions. The
rules will operate on the abstract syntax of these expressions, so we
need to be able to represent SIMPLE expressions as Ruby objects.

One way of doing this is to define a Ruby class for each distinct kind
of element from SIMPLE’s syntax—numbers, addition, multiplication,
and so on—and then represent each expression as a tree of instances of
these classes.

=============

Steps to build abstract syntax trees by hand:

$ irb --simple-prompt
>> require './add.rb'
>> require './multiply.rb'
>> require './number.rb'
>> Add.new(
*     Multiply.new(Number.new(1), Number.new(2)),
*     Multiply.new(Number.new(3), Number.new(4))
>> )
=> #<struct Add
      left=#<struct Multiply
          left=#<struct Number value=1>,
          right=#<struct Number value=2>
      >,
      right=#<struct Multiply
          left=#<struct Number value=3>,
          right=#<struct Number value=4>
      >
    >