Override #inspect on each class
The Number, Add, and Multiply classes inherit Struct’s generic
definition of #inspect, so the string representations of their
instances in the IRB console contain a lot of unimportant detail.

To make the content of an abstract syntax tree easier to see in IRB,
we’ll override #inspect on each class so that it returns a custom
string representation.

Now each abstract syntax tree will be shown in IRB as a short string of SIMPLE source
code, surrounded by «guillemets» to distinguish it from a normal Ruby
value:

$ irb --simple-prompt
>> require './add.rb'
>> require './multiply.rb'
>> require './number.rb'
>> Add.new(
*     Multiply.new(Number.new(1), Number.new(2)),
*     Multiply.new(Number.new(3), Number.new(4))
>> )
=> <1 * 2 + 3 * 4>
>> Number.new(5)
=> <5>