Implement separate #reducible? methods for Number, Add, and Multiply
Now we can begin to implement a small-step operational semantics by
defining methods that perform reductions on our abstract syntax
trees—that is, code that can take an abstract syntax tree as input and
produce a slightly reduced tree as output.

Before we can implement reduction itself, we need to be able to
distinguish expressions that can be reduced from those that
can’t.
- Add and Multiply expressions are always reducible—each of them
  represents an operation, and can be turned into a result by performing
  the calculation corresponding to that operation
- but a Number expression always represents a value, which can’t be
  reduced to anything else.

$ irb --simple-prompt
>> require './add.rb'
>> require './multiply.rb'
>> require './number.rb'
>> Number.new(1).reducible?
=> false
>> Add.new(Number.new(1), Number.new(2)).reducible?
=> true