Implement reduction for expressions
we’ll do this by defining a #reduce method for Add and
Multiply. There’s no need to define Num ber#reduce, since numbers
can’t be reduced, so we’ll just need to be careful not to call
 #reduce on an expression unless we know it’s reducible.

So what are the rules for reducing an addition expression? If the left
and right arguments are already numbers, then we can just add them
together, but what if one or both of the arguments needs reducing?
Since we’re thinking about small steps, we need to decide which
argument gets reduced first if they are both eligible for reduction.6
A common strategy is to reduce the arguments in left-to-right order,
in which case the rules will be:

• If the addition’s left argument can be reduced, reduce the left
  argument.
• If the addition’s left argument can’t be reduced but its right
  argument can, reduce the right argument.
• If neither argument can be reduced, they should both be numbers, so
  add them together.

The structure of these rules is characteristic of small-step
operational semantics. Each rule provides a pattern for the kind of
expression to which it applies—an addition with a reducible left
argument, with a reducible right argument, and with two irreducible
arguments respectively—and a description of how to build a new,
reduced expression when that pattern matches. By choosing these
particular rules, we’re specifying that a SIMPLE addition expression
uses left-to-right evaluation to reduce its arguments, as well as
deciding how those arguments should be combined once they’ve been
individually reduced.

$ irb --simple-prompt
>> require './add.rb'
>> require './multiply.rb'
>> require './number.rb'
>>  expression =
?>   Add.new(
?>    Multiply.new(Number.new(1), Number.new(2)),
?>    Multiply.new(Number.new(3), Number.new(4))
>>   )
=> <1 * 2 + 3 * 4>
>> expression.reducible?
=> true
>> expression = expression.reduce
=> «2 + 3 * 4»
>> expression.reducible?
=> true
>> expression = expression.reduce
=> «2 + 12»
>> expression.reducible?
=> true
>> expression = expression.reduce
=> «14»
>> expression.reducible?
=> false