Indroduce an environment: completes our operational semantics of expressions.
To be able to reduce a variable, we need the abstract machine to store
a mapping from variable names onto their values, an environment, as
well as the current expression.

In Ruby, we can implement this mapping as a hash, using symbols as
keys and expression objects as values; for example, the hash { x:
Number.new(2), y: Boolean.new(false) } is an environment that
associates the variables x and y with a SIMPLE number and Boolean,
respectively.

For this language, the intention is for the environment to only map
variable names onto irreducible values like Number.new(2), not onto
reducible expressions like Add.new(Number.new(1),
Number.new(2)). We’ll take care to respect this constraint later when
we write rules that change the contents of the environment.

* Variable

Given an environment, we can easily implement Variable#reduce: it just
looks up the variable’s name in the environment and returns its
value.

* Other operator

Notice that we’re now passing an environment argument into #reduce,
so we’ll need to revise the other expression classes’
implementations of #reduce to both accept and provide this argument.

* Machine

Once all the implementations of #reduce have been updated to support
environments, we also need to redefine our virtual machine to maintain
an environment and provide it to #reduce.

The machine’s definition of #run remains unchanged, but it has a new
environment attribute that is used by its new implementation of #step.

* Test

We can now perform reductions on expressions that contain variables,
as long as we also supply an environment that contains the variables’
values:

$ irb --simple-prompt
>> require './add.rb'
>> require './multiply.rb'
>> require './number.rb'
>> require './boolean.rb'
>> require './less_than.rb'
>> require './variable.rb'
>> require './machine.rb'
>> Machine.new(
?>  Add.new(Variable.new(:x), Variable.new(:y)),
?>  { x: Number.new(3), y: Number.new(4) }
>>  ).run
x + y
3 + y
3 + 4
7
=> nil

* Conclusion

The introduction of an environment completes our operational semantics
of expressions. We’ve designed an abstract machine that begins with
an initial expression and environment, and then uses the current
expression and environment to produce a new expression in each small
reduction step, leaving the environment unchanged.