The simplest possible statement: do nothing
The simplest possible statement is one that does nothing: it can’t be
reduced, so it can’t have any effect on the environment.

1. All of our other syntax classes inherit from a Struct class, but
   DoNothing doesn’t inherit from anything. This is because DoNothing
   doesn’t have any attributes, and unfortunately, Struct.new doesn’t
   let us pass an empty list of attribute names.

2. We want to be able to compare any two statements to see if they’re
   equal. The other syntax classes inherit an implementation of #== from
   Struct, but DoNothing has to define its own.

A statement that does nothing might seem pointless, but it’s
convenient to have a special statement that represents a program whose
execution has completed successfully. We’ll arrange for other
statements to eventually reduce to «do-nothing» once they’ve finished
doing their work.