Implement an Assign class
An assignment statement consists of a variable name (x), an equals
symbol, and an expression («x + 1»). If the expression within the
assignment is reducible, we can just reduce it according to the
expression reduction rules and end up with a new assignment statement
containing the reduced expression.

For example, reducing «x = x + 1» in an environment where the variable
x has the value «2» should leave us with the statement «x = 2 + 1»,
and reducing it again should produce «x = 3».

But then what? If the expression is already a value like «3», then we
should just perform the assignment, which means updating the
environment to associate that value with the appropriate variable
name. So reducing a statement needs to produce not just a new, reduced
statement but also a new environment, which will sometimes be
different from the environment in which the reduction was performed.

So reducing «x = 3» in an empty environment should produce the new
environment { x: Number.new(3) }, but we also expect the statement to
be reduced somehow; otherwise, our abstract machine will keep
assigning «3» to x forever. That’s what «donothing» is for: a
completed assignment reduces to «do-nothing», indicating that
reduction of the statement has finished and that whatever’s in the
new environment may be considered its result.

To summarize, the reduction rules for assignment are:

• If the assignment’s expression can be reduced, then reduce it,
  resulting in a reduced assignment statement and an unchanged
  environment.
• If the assignment’s expression can’t be reduced, then update the
  environment to associate that expression with the assignment’s
  variable, resulting in a «do-noth ing» statement and a new
  environment.