Big-step evaluation of an assignment statement
We can think of bigstep statement evaluation as a process that always
turns a statement and an initial environment into a final environment,
avoiding the small-step complication of also having to deal with the
intermediate statement generated by #reduce.

Big-step evaluation of an assignment statement, for example, should
fully evaluate its expression and return an updated environment
containing the resulting value.