Big-step evaluation of sequence statements
For sequences, we just need to evaluate both statements, but the
initial environment needs to be “threaded through” these two
evaluations, so that the result of evaluating the first statement
becomes the environment in which the second statement is
evaluated.

This can be written in Ruby by using the first evaluation’s result as
the argument to the second.

$ irb --simple-prompt
>> require './number.rb'
>> require './boolean.rb'
>> require './variable.rb'
>> require './add.rb'
>> require './multiply.rb'
>> require './less_than.rb'
>> require './do_nothing.rb'
>> require './assign.rb'
>> require './if.rb'
>> require './sequence.rb'

>> statement =
?>  Sequence.new(
?>   Assign.new(:x, Add.new(Number.new(1), Number.new(1))),
?>   Assign.new(:y, Add.new(Variable.new(:x), Number.new(3)))
>>  )
=> <x = 1 + 1; y = x + 3>
>> statement.evaluate({})
=> {:x=><2>, :y=><5>}