Write implementations of a #to_ruby method for the Number and Boolean classes
Let’s try giving a denotational semantics for SIMPLE by translating
it into Ruby. In practice, this means turning an abstract syntax tree
into a string of Ruby code that somehow captures the intended meaning
of that syntax.

But what is the “intended meaning”? What should Ruby denotations of
our expressions and statements look like? We’ve already seen
operationally that an expression takes an environment and turns it
into a value;

One way to express this in Ruby is with a proc that takes some
argument representing an environment argument and returns some Ruby
object representing a value. For simple constant expressions like «5»
and «false», we won’t be using the environment at all, so we only
need to worry about how their eventual result can be represented as a
Ruby object. Fortunately, Ruby already has objects specifically
designed to represent these values: we can use the Ruby value 5 as the
result of the SIMPLE expression «5», and likewise, the Ruby value
false as the result of «false».

$ irb --simple-prompt
>> require './number.rb'
>> require './boolean.rb'

>> Number.new(5).to_ruby
=> "-> e { 5 }"

>> Boolean.new(false).to_ruby
=> "-> e { false }"

Each of these methods produces a string that happens to contain Ruby
code, and because Ruby is a language whose meaning we already
understand, we can see that both of these strings are programs that
build procs. Each proc takes an environment argument called e,
completely ignores it, and returns a Ruby value.

Because these denotations are strings of Ruby source code, we can
check their behavior in IRB by using Kernel#eval to turn them into
real, callable Proc objects.

>> proc = eval(Number.new(5).to_ruby)
=> #<Proc (lambda)>
>> proc.call({})
=> 5

>> proc = eval(Boolean.new(false).to_ruby)
=> #<Proc (lambda)>
>> proc.call({})
=> false

Chapter 1.2.3 Proc

A proc is an unevaluated chunk of Ruby code that can be passed around
and evaluated on demand; other languages call this an “anonymous
function” or “lambda.” There are several ways of writing a proc
literal, the most compact of which is the -> argu ments { body }
syntax:

>> multiply = -> x, y { x * y }
=> #<Proc (lambda)>

>> multiply.call(6, 9)
=> 54

>> multiply.call(2, 3)
=> 6

As well as the .call syntax, procs can be called by using square brackets:

>> multiply[3, 4]
=> 12