Implement Variable#to_ruby
To denote expressions that do use the environment, we need to decide
how environments are going to be represented in Ruby. We’ve already
seen environments in our operational semantics, and since they were
implemented in Ruby, we can just reuse our earlier idea of
representing an environment as a hash. The details will need to
change, though, so beware the subtle difference:

- in our operational semantics, the environment lived inside the
  virtual machine and associated variable names with SIMPLE abstract
  syntax trees like Number.new(5)
- but in our denotational semantics, the environment exists in the
  language we’re translating our programs into, so it needs to make
  sense in that world instead of the “outside world” of a virtual
  machine.

In particular, this means that our denotational environments should
associate variable names with native Ruby values like 5 rather than
with objects representing SIMPLE syntax. We can think of an
operational environment like { x: Number.new(5) } as having a
denotation of '{ x: 5 }' in the language we’re translating into, and
we just need to keep our heads straight because both the
implementation metalanguage and the denotation language happen to be
Ruby.

$ irb --simple-prompt
>> require './variable.rb'

>> expression = Variable.new(:x)
=> «x»
>> expression.to_ruby
=> "-> e { e[:x] }"

>> proc = eval(expression.to_ruby)
=> #<Proc (lambda)>
>> proc.call({ x: 7 })
=> 7