Move onto denoting larger expressions like Add, Multiply, and LessThan
An important aspect of denotational semantics is that it’s
compositional: the denotation of a program is constructed from the
denotations of its parts.

Here we’re using string concatenation to compose the denotation of an
expression out of the denotations of its subexpressions. We know that
each subexpression will be denoted by a proc’s Ruby source, so we can
use them as part of a larger piece of Ruby source that calls those
procs with the supplied environment and does some computation with
their return values. Here’s what the resulting denotations look like:

$ irb --simple-prompt
>> require './number.rb'
>> require './boolean.rb'
>> require './variable.rb'
>> require './add.rb'
>> require './multiply.rb'
>> require './less_than.rb'

>> Add.new(Variable.new(:x), Number.new(1)).to_ruby
=> "-> e { (-> e { e[:x] }).call(e) + (-> e { 1 }).call(e) }"

>> LessThan.new(Add.new(Variable.new(:x), Number.new(1)), Number.new(3)).to_ruby
=> "-> e { (-> e { (-> e { e[:x] }).call(e) + (-> e { 1 }).call(e)
}).call(e) < (-> e { 3 }).call(e) }"

These denotations are now complicated enough that it’s difficult to
see whether they do the right thing. Let’s try them out to make sure:

>> environment = { x: 3 }
=> {:x=>3}
>> proc = eval(Add.new(Variable.new(:x), Number.new(1)).to_ruby)
=> #<Proc (lambda)>
>> proc.call(environment)
=> 4
>> proc = eval(
 LessThan.new(Add.new(Variable.new(:x), Number.new(1)), Number.new(3)).to_ruby
 )
=> #<Proc (lambda)>
>> proc.call(environment)
=> false