Implement a collection of rules, which is called a rulebook
Deterministic finite automata are intended as abstract models of
computation. We’ve drawn diagrams of a couple of example machines and
thought about their behavior, but these machines don’t physically
exist, so we can’t literally feed them input and see how they
behave.

Fortunately DFAs are so simple that we can easily build a
simulation of one in Ruby and interact with it directly.

Let’s begin building that simulation by implementing a collection of
rules, which we’ll call a rulebook.

This code establishes a simple API for rules: each rule has
- an #applies_to? method, which returns true or false to indicate whether
that rule applies in a particular situation
- and a #follow method that returns information about how the machine
should change when a rule is followed.

DFARulebook#next_state uses these methods to locate the
correct rule and discover what the next state of the DFA should be.

A rulebook lets us wrap up many rules into a single object and ask it
questions about which state comes next:

$ irb --simple-prompt
>> require './dfa_rulebook.rb'
>> require './fa_rule.rb'

>> rulebook = DFARulebook.new([
 FARule.new(1, 'a', 2), FARule.new(1, 'b', 1),
 FARule.new(2, 'a', 2), FARule.new(2, 'b', 3),
 FARule.new(3, 'a', 3), FARule.new(3, 'b', 3)
 ])
=> #<struct DFARulebook …>

b --a--> a --b--> a
1        2        3
                  b

>> rulebook.next_state(1, 'a')
=> 2
>> rulebook.next_state(1, 'b')
=> 1
>> rulebook.next_state(2, 'b')
=> 3