Create DFADesign to automatically build one-off instances
Once a DFA object has been fed some input, it’s probably not in its
start state anymore, so we can’t reliably reuse it to check a
completely new sequence of inputs. That means we have to recreate it
from scratch—using the same start state, accept states, and rulebook
as before—every time we want to see whether it will accept a new
string. We can avoid doing this manually by wrapping up its
constructor’s arguments in an object that represents the design of a
particular DFA and relying on that object to automatically build
one-off instances of that DFA whenever we want to check for acceptance
of a string.

The #tap method evaluates a block and then returns the object it was
called on.

DFADesign#accepts? uses the DFADesign#to_dfa method to create a new
instance of DFA and then calls #read_string? to put it into an
accepting or rejecting state.

>> require './dfa_design.rb'

>> dfa_design = DFADesign.new(1, [3], rulebook)
=> #<struct DFADesign …>
>> dfa_design.accepts?('a')
=> false
>> dfa_design.accepts?('baa')
=> false
>> dfa_design.accepts?('baba')
=> true