The implementation of NFARulebook
There’s an easy way to simulate an NFA without needing to rewind our
progress, spawn threads, or know all the input characters in advance.

In fact, just as we simulated a single DFA by keeping track
of its current state, we can simulate a single NFA by keeping track of
all its possible current states. This is simpler and more efficient
than simulating multiple NFAs that go off in different directions, and
it turns out to achieve the same thing in the end. If we did simulate
many separate machines, then all we’d care about is what state each
of them was in, but any machines in the same state are completely
indistinguishable, so we don’t lose anything by collapsing all those
possibilities down into a single machine and asking “which states
could it be in by now?” instead.

First we need a rulebook suitable for storing an NFA’s rules.

- A DFA rulebook always returns a single state when we ask it where
  the DFA should go next after reading a particular character while in
  a specific state,
- but an NFA rulebook needs to answer a different question: when an
  NFA is possibly in one of several states, and it reads a particular
  character, what states can it possibly be in next?

$ irb --simple-prompt
>> require './nfa_rulebook.rb'
>> require './fa_rule.rb'

>> rulebook = NFARulebook.new([
 FARule.new(1, 'a', 1), FARule.new(1, 'b', 1), FARule.new(1, 'b', 2),
 FARule.new(2, 'a', 3), FARule.new(2, 'b', 3),
 FARule.new(3, 'a', 4), FARule.new(3, 'b', 4)
 ])
=> #<struct NFARulebook rules=[…]>

>> rulebook.next_states(Set[1], 'b')
=> #<Set: {1, 2}>
>> rulebook.next_states(Set[1, 2], 'a')
=> #<Set: {1, 3}>
>> rulebook.next_states(Set[1, 3], 'b')
=> #<Set: {1, 2, 4}>