Implement an NFA class to represent the simulated machine
NFA#accepting? works by checking whether there are any states in the
intersection between current_states and accept_states—that is, whether
any of the possible current states is also one of the accept states.

This NFA class is very similar to our DFA class from earlier. The
difference is that it has a set of possible current_states instead of
a single definite current_state, so it’ll say it’s in an accept
state if any of its current_states is an accept state:

$ irb --simple-prompt
>> require './nfa_rulebook.rb'
>> require './fa_rule.rb'
>> require './nfa.rb'

>> NFA.new(Set[1], [4], rulebook).accepting?
=> false
>> NFA.new(Set[1, 2, 4], [4], rulebook).accepting?
=> true

As with the DFA class, we can implement a #read_character method for
reading a single character of input, and a #read_string method for
reading several in sequence.

These methods really are almost identical to their DFA counterparts;
we’re just saying current_states and next_states in #read_character
instead of current_state and next_state.

Now we’re able to start a simulated NFA, pass characters in, and ask
whether the input so far has been accepted:

>> nfa = NFA.new(Set[1], [4], rulebook); nfa.accepting?
=> false
>> nfa.read_character('b'); nfa.accepting?
=> false
>> nfa.read_character('a'); nfa.accepting?
=> false
>> nfa.read_character('b'); nfa.accepting?
=> true

>> nfa = NFA.new(Set[1], [4], rulebook)
=> #<struct NFA current_states=#<Set: {1}>, accept_states=[4], rulebook=…>
>> nfa.accepting?
=> false
>> nfa.read_string('bbbbb'); nfa.accepting?
=> true