Implement #to_nfa_design methods for Empty and Literal to generate NFA
Now that we have a way of representing the syntax of a regular
expression as a tree of Ruby objects, how can we convert that syntax
into an NFA?

We need to decide how instances of each syntax class should be turned
into NFAs. The easiest class to convert is Empty, which we should
always turn into the one-state NFA that only accepts the empty string.

Similarly, we should turn any literal, single-character pattern into
the NFA that only accepts the single-character string containing that
character.

We can check that the NFAs generated from Empty and Literal regular
expressions accept the strings we want them to:

$ irb --simple-prompt
>> require './nfa_rulebook.rb'
>> require './fa_rule.rb'
>> require './nfa.rb'
>> require './nfa_design.rb'

>> require './pattern.rb'
>> require './empty.rb'
>> require './literal.rb'

>> nfa_design = Empty.new.to_nfa_design
=> #<struct NFADesign …>
>> nfa_design.accepts?('')
=> true
>> nfa_design.accepts?('a')
=> false
>> nfa_design = Literal.new('a').to_nfa_design
=> #<struct NFADesign …>
>> nfa_design.accepts?('')
=> false
>> nfa_design.accepts?('a')
=> true
>> nfa_design.accepts?('b')
=> false