Turn Concatenate regular expressions into NFAs
If we have two regular expressions that we already know how to turn
into NFAs, how can we build an NFA to represent the concatenation of
those regular expressions? For example, given that we can turn the
single-character regular expressions a and b into NFAs, how do we turn
ab into one?

In the ab case, we can connect the two NFAs in sequence, joining them
together with a free move, and keeping the second NFA’s accept state:

-->start--a-->end       -->start--b-->end
                \       /
                 ---To--
                    |
  -->start--a-->end--nil-->start--b-->end

This technique works in other cases too. Any two NFAs can be
concatenated by turning every accept state from the first NFA into a
nonaccept state and connecting it to the start state of the second NFA
with a free move. Once the concatenated machine has read a sequence of
inputs that would have put the first NFA into an accept state, it can
spontaneously move into a state corresponding to the start state of
the second NFA, and then reach an accept state by reading a sequence
of inputs that the second NFA would have accepted.

So, the raw ingredients for the combined machine are:

• The start state of the first NFA
• The accept states of the second NFA
• All the rules from both NFAs
• Some extra free moves to connect each of the first NFA’s old accept
  states to the second NFA’s old start state

We can turn this idea into an implementation of Concatenate#to_nfa_design.

This code first converts the first and second regular expressions into
NFADesigns, then combines their states and rules in the appropriate
way to make a new NFADesign. It works as expected for the simple ab
case:

$ irb --simple-prompt
>> require './nfa_rulebook.rb'
>> require './fa_rule.rb'
>> require './nfa.rb'
>> require './nfa_design.rb'

>> require './pattern.rb'
>> require './concatenate.rb'

>> pattern = Concatenate.new(Literal.new('a'), Literal.new('b'))
=> /ab/
>> pattern.matches?('a')
=> false
>> pattern.matches?('ab')
=> true
>> pattern.matches?('abc')
=> false

This conversion process is recursive—Concatenate#to_nfa_design calls
 #to_nfa_design on other objects—so it also works for more deeply
 #nested cases like the regular expression abc, which contains two
 #concatenations (a concatenated with b concatenated with c):

>> pattern =
 Concatenate.new(
   Literal.new('a'),
   Concatenate.new(Literal.new('b'), Literal.new('c'))
 )
=> /abc/
>> pattern.matches?('a')
=> false
>> pattern.matches?('ab')
=> false
>> pattern.matches?('abc')
=> true