Convert a Choose expression into an NFA
In the simplest case, the NFAs for the regular expressions a and b can
be combined to build an NFA for the regular expression a|b by adding a
new start state and using free moves to connect it to the previous
start states of the two original machines:

-->start--a-->end        -->start--b-->end
        |  <------------|
        | To

        --nil-->start--a-->end
-->start
        --nil-->start--b-->end

Before the a|b NFA has read any input, it can use a free move to go
into either of the original machines’ start states, from which point
it can read either 'a' or 'b' to reach an accept state. Again, it’s
just as easy to glue together any two machines by adding a new start
state and two free moves.

In this case, the ingredients for the combined machine are:
• A new start state
• All the accept states from both NFAs
• All the rules from both NFAs
• Two extra free moves to connect the new start state to each of the
  NFA’s old start states

Again, this is easy to implement as Choose#to_nfa_design.

The implementation works nicely:

$ irb --simple-prompt
>> require './nfa_rulebook.rb'
>> require './fa_rule.rb'
>> require './nfa.rb'
>> require './nfa_design.rb'

>> require './pattern.rb'
>> require './choose.rb'
>> require './literal.rb'

>> pattern = Choose.new(Literal.new('a'), Literal.new('b'))
=> /a|b/
>> pattern.matches?('a')
=> true
>> pattern.matches?('b')
=> true
>> pattern.matches?('c')
=> false