Convert a Repeat expression into an NFA
And finally, repetition: how can we turn an NFA that matches a string
exactly once into an NFA that matches the same string zero or more
times? We can build an NFA for a* by starting with the NFA for a and
making two additions:

• Add a free move from its accept state to its start state, so it can
  match more than   one 'a'.

• Add a new accepting start state with a free move to the old start
  state, so it can   match the empty string.

Here’s how that looks:

-->start--a-->end     (zero or more times)
     |
     |To
     |
-->end--nil-->x--a-->end
              <--nil--

The free move from the old accept state to the old start state allows
the machine to match several times instead of just once ('aa', 'aaa',
etc.), and the new start state allows it to match the empty string
without affecting what other strings it can accept.5 We can do the
same for any NFA as long as we connect each old accept state to the
old start state with a free move.

This time we need:

• A new start state, which is also an accept state
• All the accept states from the old NFA
• All the rules from the old NFA
• Some extra free moves to connect each of the old NFA’s accept
  states to its old start state
• Another extra free move to connect the new start state to the old start state

And check that it works:

$ irb --simple-prompt
>> require './nfa_rulebook.rb'
>> require './fa_rule.rb'
>> require './nfa.rb'
>> require './nfa_design.rb'

>> require './pattern.rb'
>> require './literal.rb'
>> require './repeat.rb'

>> pattern = Repeat.new(Literal.new('a'))
=> /a*/
>> pattern.matches?('')
=> true
>> pattern.matches?('a')
=> true
>> pattern.matches?('aaaa')
=> true
>> pattern.matches?('b')
=> false

Now that we have #to_nfa_design implementations for each class of
regular expression syntax, we can build up complex patterns and use
them to match strings:

$ irb --simple-prompt
>> require './nfa_rulebook.rb'
>> require './fa_rule.rb'
>> require './nfa.rb'
>> require './nfa_design.rb'

>> require './pattern.rb'
>> require './choose.rb'
>> require './concatenate.rb'
>> require './empty.rb'
>> require './literal.rb'
>> require './repeat.rb'

>> pattern =
 Repeat.new(
   Concatenate.new(
     Literal.new('a'),
     Choose.new(Empty.new, Literal.new('b'))
   )
 )
=> /(a(|b))*/
>> pattern.matches?('')
=> true
>> pattern.matches?('a')
=> true
>> pattern.matches?('ab')
=> true
>> pattern.matches?('aba')
=> true
>> pattern.matches?('abab')
=> true
>> pattern.matches?('abaab')
=> true
>> pattern.matches?('abba')
=> false