Create a parser for pattern syntax
We’ve almost built a complete (albeit basic) regular expression
implementation. The only missing piece is a parser for pattern syntax:
it would be much more convenient if we could just write (a(|b))*
instead of building the abstract syntax tree manually with
Repeat.new(Concatenate.new(Literal.new('a'), Choose.new(Empty.new,
Literal.new ('b')))). We saw in “Implementing Parsers” on page 58
that it’s not difficult to use Treetop to generate a parser that can
automatically transform raw syntax into an AST, so let’s do that here
to finish off our implementation.

The order of rules reflects the precedence of each operator: the |
operator binds loosest, so the choose rule goes first, with the higher
precedence operator rules appearing farther down the grammar.

Now we have all the pieces we need to parse a regular expression, turn
it into an abstract syntax tree, and use it to match strings:

$ irb --simple-prompt
>> require './nfa_rulebook.rb'
>> require './fa_rule.rb'
>> require './nfa.rb'
>> require './nfa_design.rb'

>> require './pattern.rb'
>> require './choose.rb'
>> require './concatenate.rb'
>> require './empty.rb'
>> require './literal.rb'
>> require './repeat.rb'

>> require 'treetop'
=> true
>> Treetop.load('pattern')
=> PatternParser
>> parse_tree = PatternParser.new.parse('(a(|b))*')
=> SyntaxNode+Repeat1+Repeat0 offset=0, "(a(|b))*" (to_ast,brackets):
  SyntaxNode+Brackets1+Brackets0 offset=0, "(a(|b))" (to_ast,choose):
    SyntaxNode offset=0, "("
    SyntaxNode+Concatenate1+Concatenate0 offset=1, "a(|b)" (to_ast,first,rest):
      SyntaxNode+Literal0 offset=1, "a" (to_ast)
      SyntaxNode+Brackets1+Brackets0 offset=2, "(|b)" (to_ast,choose):
        SyntaxNode offset=2, "("
        SyntaxNode+Choose1+Choose0 offset=3, "|b" (to_ast,first,rest):
          SyntaxNode+Empty0 offset=3, "" (to_ast)
          SyntaxNode offset=3, "|"
          SyntaxNode+Literal0 offset=4, "b" (to_ast)
        SyntaxNode offset=5, ")"
    SyntaxNode offset=6, ")"
  SyntaxNode offset=7, "*"
>> pattern = parse_tree.to_ast
=> /(a(|b))*/
>> pattern.matches?('abaab')
=> true
>> pattern.matches?('abba')
=> false