Add an optional “current states” parameter to NFADesign#to_nfa
add an optional “current states” parameter to NFADesign#to_nfa so
that we can build an NFA instance with any set of current states, not
just the NFADesign’s start state.

Previously, the simulation of an NFA could only begin in its start
state, but this new parameter gives us a way of jumping in at any
other point:

$ irb --simple-prompt
>> require './fa_rule.rb'
>> require './dfa_rulebook.rb'
>> require './dfa.rb'
>> require './dfa_design.rb'
>> require './nfa_rulebook.rb'
>> require './nfa.rb'
>> require './nfa_design.rb'

>> rulebook = NFARulebook.new([
?>    FARule.new(1, 'a', 1), FARule.new(1, 'a', 2), FARule.new(1, nil, 2),
?>    FARule.new(2, 'b', 3),
?>    FARule.new(3, 'b', 1), FARule.new(3, nil, 2)
>> ])
=> #<struct NFARulebook rules=[#<FARule 1 -- a --> 1>, #<FARule 1 -- a --> 2>, #
<FARule 1 --  --> 2>, #<FARule 2 -- b --> 3>, #<FARule 3 -- b --> 1>, #<FARule 3
 --  --> 2>]>
>> nfa_design = NFADesign.new(1, [3], rulebook)
=> #<struct NFADesign start_state=1, accept_states=[3], rulebook=#<struct NFARul
ebook rules=[#<FARule 1 -- a --> 1>, #<FARule 1 -- a --> 2>, #<FARule 1 --  -->
2>, #<FARule 2 -- b --> 3>, #<FARule 3 -- b --> 1>, #<FARule 3 --  --> 2>]>>
>> nfa_design.to_nfa.current_states
=> #<Set: {1, 2}>
>> nfa_design.to_nfa(Set[2]).current_states
=> #<Set: {2}>
>> nfa_design.to_nfa(Set[3]).current_states
=> #<Set: {3, 2}>

Now we can create an NFA in any set of possible states, feed it a
character, and see what states it might end up in, which is a crucial
step for converting an NFA into a DFA. When our NFA is in state 2 or 3
and reads a b, what states can it be in afterward?

>> nfa = nfa_design.to_nfa(Set[2, 3])
=> #<struct NFA current_states=#<Set: {2, 3}>, accept_states=[3], rulebook=#<str
uct NFARulebook rules=[#<FARule 1 -- a --> 1>, #<FARule 1 -- a --> 2>, #<FARule
1 --  --> 2>, #<FARule 2 -- b --> 3>, #<FARule 3 -- b --> 1>, #<FARule 3 --  -->
 2>]>>
>> nfa.read_character('b'); nfa.current_states
=> #<Set: {3, 1, 2}>

The answer: state 1, 2, or 3, as we already discovered during the
manual conversion process. (Remember that the order of elements in a
Set doesn’t matter.)