Create the NFASimulation class
Create the NFASimulation class and giving it a method to calculate how
the state of the simulation will change in response to a particular
input. We’re thinking of the state of the simulation as being the
current set of possible states for the NFA (e.g., “1, 2, or 3”), so
we can write a #next_state method that takes a simulation state and a
character, feeds that character to an NFA corresponding to that state,
and gets a new state back out by inspecting the NFA afterward.

This gives us a convenient way to explore the different states of the
simulation:

>>  require './nfa_simulation.rb'

>> simulation = NFASimulation.new(nfa_design)
=> #<struct NFASimulation nfa_design=#<struct NFADesign start_state=1, accept_st
ates=[3], rulebook=#<struct NFARulebook rules=[#<FARule 1 -- a --> 1>, #<FARule
1 -- a --> 2>, #<FARule 1 --  --> 2>, #<FARule 2 -- b --> 3>, #<FARule 3 -- b --
> 1>, #<FARule 3 --  --> 2>]>>>
>> simulation.next_state(Set[1, 2], 'a')
=> #<Set: {1, 2}>
>> simulation.next_state(Set[1, 2], 'b')
=> #<Set: {3, 2}>
>> simulation.next_state(Set[3, 2], 'b')
=> #<Set: {1, 3, 2}>
>> simulation.next_state(Set[1, 3, 2], 'b')
=> #<Set: {1, 3, 2}>
>> simulation.next_state(Set[1, 3, 2], 'a')
=> #<Set: {1, 2}>