Systematically explore the simulation states
Now we need a way to systematically explore the simulation states and
record our discoveries as the states and rules of a DFA. We intend to
use each simulation state directly as a DFA state, so the first step
is to implement NFASimulation#rules_for, which builds all the rules
leading from a particular simulation state by using #next_state to
discover the destination of each rule. “All the rules” means a rule
for each possible input character, so we also define an
NFARulebook#alphabet helper method to tell us what characters the
original NFA can read.

As intended, this lets us see how different inputs will take the
simulation between different states:

$ irb --simple-prompt
>> require './fa_rule.rb'
>> require './dfa_rulebook.rb'
>> require './dfa.rb'
>> require './dfa_design.rb'
>> require './nfa_rulebook.rb'
>> require './nfa.rb'
>> require './nfa_design.rb'
>> require './nfa_simulation.rb'

>> rulebook = NFARulebook.new([
?>  FARule.new(1, 'a', 1), FARule.new(1, 'a', 2), FARule.new(1, nil, 2),
?>  FARule.new(2, 'b', 3),
?>  FARule.new(3, 'b', 1), FARule.new(3, nil, 2)
>> ])
>> nfa_design = NFADesign.new(1, [3], rulebook)
>> simulation = NFASimulation.new(nfa_design)

>> rulebook.alphabet
=> ["a", "b"]
>> simulation.rules_for(Set[1, 2])
=> [
 #<FARule #<Set: {1, 2}> --a--> #<Set: {1, 2}>>,
 #<FARule #<Set: {1, 2}> --b--> #<Set: {3, 2}>>
 ]
>> simulation.rules_for(Set[3, 2])
=> [
 #<FARule #<Set: {3, 2}> --a--> #<Set: {}>>,
 #<FARule #<Set: {3, 2}> --b--> #<Set: {1, 3, 2}>>
 ]