Find more states and rules
The #rules_for method gives us a way of exploring outward from a known
simulation state and discovering new ones, and by doing this
repeatedly, we can find all possible simulation states. We can do this
with an NFASimulation#discover_states_and_rules method, which
recursively finds more states in a similar way to NFARulebook#follow_free_moves.

Initially, we only know about a single state of the simulation: the
set of possible states of our NFA when we put it into its start
state. #discover_states_and_rules explores outward from this starting
point, eventually finding all four states and eight rules of the
simulation:

$ irb --simple-prompt
>> require './fa_rule.rb'
>> require './dfa_rulebook.rb'
>> require './dfa.rb'
>> require './dfa_design.rb'
>> require './nfa_rulebook.rb'
>> require './nfa.rb'
>> require './nfa_design.rb'
>> require './nfa_simulation.rb'

>> rulebook = NFARulebook.new([
?>  FARule.new(1, 'a', 1), FARule.new(1, 'a', 2), FARule.new(1, nil, 2),
?>  FARule.new(2, 'b', 3),
?>  FARule.new(3, 'b', 1), FARule.new(3, nil, 2)
>> ])
>> nfa_design = NFADesign.new(1, [3], rulebook)
>> simulation = NFASimulation.new(nfa_design)

>> start_state = nfa_design.to_nfa.current_states
=> #<Set: {1, 2}>
>> simulation.discover_states_and_rules(Set[start_state])
=> [
     #<Set: {
       #<Set: {1, 2}>,
       #<Set: {3, 2}>,
       #<Set: {}>,
       #<Set: {1, 3, 2}>
     }>,
     [
       #<FARule #<Set: {1, 2}> --a--> #<Set: {1, 2}>>,
       #<FARule #<Set: {1, 2}> --b--> #<Set: {3, 2}>>,
       #<FARule #<Set: {3, 2}> --a--> #<Set: {}>>,
       #<FARule #<Set: {3, 2}> --b--> #<Set: {1, 3, 2}>>,
       #<FARule #<Set: {}> --a--> #<Set: {}>>,
       #<FARule #<Set: {}> --b--> #<Set: {}>>,
       #<FARule #<Set: {1, 3, 2}> --a--> #<Set: {1, 2}>>,
       #<FARule #<Set: {1, 3, 2}> --b--> #<Set: {1, 3, 2}>>
     ]
 ]

The final thing we need to know for each simulation state is whether
it should be treated as an accept state, but that’s easy to check by
asking the NFA at that point in the simulation:

>> nfa_design.to_nfa(Set[1, 2]).accepting?
=> false
>> nfa_design.to_nfa(Set[2, 3]).accepting?
=> true