Implement a Stack class
This is a purely functional stack. The #push and #pop methods are
nondestructive: they each return a new Stack instance rather than
modifying the existing one. Creating a new object every time makes
this implementation less efficient than a conventional stack with
destructive #push and #pop operations (and we could just use Array
directly if we wanted that) but also makes it easier to work with,
because we don’t need to worry about the consequences of modifying a
Stack that’s being used in more than one place.

A Stack object stores its contents in an underlying array and exposes
simple #push and #pop operations to push characters onto the stack and
pop them off, #plus a #top operation to read the character at the top
of the stack:

>> irb --simple-prompt
>> require './stack.rb'

>> stack = Stack.new(['a', 'b', 'c', 'd', 'e'])
=> #<Stack (a)bcde>
>> stack.top
=> "a"
>> stack.pop.pop.top
=> "c"
>> stack.push('x').push('y').top
=> "y"
>> stack.push('x').push('y').pop.top
=> "x"